requirements.txt
# LangChain core
langchain>=0.1.0
langchain-core>=0.1.0

# LLM providers
langchain-gigachat>=0.1.0
langchain-openai>=0.0.5

# Agent framework
langgraph>=0.0.20

# Observability
rich>=13.0

# Utilities
pyyaml>=6.0

# Optional: MLflow (uncomment if needed)
# mlflow>=2.10

# UI (опционально — только для streamlit_ui.py)
streamlit>=1.35

# Визуализация (используется инструментом plot_chart)
matplotlib>=3.8
